Improve the RUN speed of the Dockerfile
RUN incurs quite an overhead because it makes a cache entry each
time it's called. There's no reason to do so as only one file is
edited that never makes sense to have cached partially. We rewrite
the RUN statements to a single statement so that less intermediate
containers are cached.