Enable zend assertions in development and ci
We use `assert` in places to satisfy PHPStan when we know things should be a certain way.
Similarly we use it in tests when we know things should be a certain way (or to actually make
test assertions).

However, until now `assert` wasn't actually executed in our local environments or in our test
runners which can cause tests to fail unpredictably elsewhere.

This adds `zend.assertions = 1` to enable in ci mode. We add a duplicate line so
that the top of the `php.ini` is still the same as the production one and it's clear that it's a
dev- and ci-only override.